requirements.txt
pytest==7.4.3
selene==2.0.0rc4
allure-pytest==2.13.2
pydantic-settings==2.0.3